Temporary roots
In the game I’m working on I would like to show modal popups at a place where there’s no underlying gui to attach it to. I thought it would be nice if PopupManager could create a temporary root in those cases.

By passing the stage to config.root a root will be created with the stage as parent and then be destroyed when the popup is disposed.